producer/requirements.txt
kafka-python
fastapi